# src/api/endpoint/models/info.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from .base_model_ import Model
from . import util
from .version import Version


class Info(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, name: str = None, version: Version = None):  # noqa: E501
        """Info - a model defined in Swagger

        :param name: The name of this Info.  # noqa: E501
        :type name: str
        :param version: The version of this Info.  # noqa: E501
        :type version: Version
        """
        self.swagger_types = {
            'name': str,
            'version': Version
        }

        self.attribute_map = {
            'name': 'name',
            'version': 'version'
        }

        self._name = name
        self._version = version

    @classmethod
    def from_dict(cls, dikt) -> 'Info':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Info of this Info.  # noqa: E501
        :rtype: Info
        """
        return util.deserialize_model(dikt, cls)

    @property
    def name(self) -> str:
        """Gets the name of this Info.


        :return: The name of this Info.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Info.


        :param name: The name of this Info.
        :type name: str
        """

        self._name = name

    @property
    def version(self) -> Version:
        """Gets the version of this Info.


        :return: The version of this Info.
        :rtype: Version
        """
        return self._version

    @version.setter
    def version(self, version: Version):
        """Sets the version of this Info.


        :param version: The version of this Info.
        :type version: Version
        """

        self._version = version