# src/api/endpoint/models/switchable_device_datasheet.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from .base_model_ import Model
from . import util
from .switchable_device_datasheet_datasheet import SwitchableDeviceDatasheetDatasheet


class SwitchableDeviceDatasheet(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, datasheet: SwitchableDeviceDatasheetDatasheet = None):  # noqa: E501
        """SwitchableDeviceDatasheet - a model defined in Swagger

        :param datasheet: The datasheet of this SwitchableDeviceDatasheet.  # noqa: E501
        :type datasheet: SwitchableDeviceDatasheetDatasheet
        """
        self.swagger_types = {
            'datasheet': SwitchableDeviceDatasheetDatasheet
        }

        self.attribute_map = {
            'datasheet': 'datasheet'
        }

        self._datasheet = datasheet

    @classmethod
    def from_dict(cls, dikt) -> 'SwitchableDeviceDatasheet':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SwitchableDeviceDatasheet of this SwitchableDeviceDatasheet.  # noqa: E501
        :rtype: SwitchableDeviceDatasheet
        """
        return util.deserialize_model(dikt, cls)

    @property
    def datasheet(self) -> SwitchableDeviceDatasheetDatasheet:
        """Gets the datasheet of this SwitchableDeviceDatasheet.


        :return: The datasheet of this SwitchableDeviceDatasheet.
        :rtype: SwitchableDeviceDatasheetDatasheet
        """
        return self._datasheet

    @datasheet.setter
    def datasheet(self, datasheet: SwitchableDeviceDatasheetDatasheet):
        """Sets the datasheet of this SwitchableDeviceDatasheet.


        :param datasheet: The datasheet of this SwitchableDeviceDatasheet.
        :type datasheet: SwitchableDeviceDatasheetDatasheet
        """

        self._datasheet = datasheet