# src/api/endpoint/models/task.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from .base_model_ import Model
from . import util
from .task_description import TaskDescription


class Task(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, description: TaskDescription = None, status: str = None):  # noqa: E501
        """Task - a model defined in Swagger

        :param description: The description of this Task.  # noqa: E501
        :type description: TaskDescription
        :param status: The status of this Task.  # noqa: E501
        :type status: str
        """
        self.swagger_types = {
            'description': TaskDescription,
            'status': str
        }

        self.attribute_map = {
            'description': 'description',
            'status': 'status'
        }

        self._description = description
        self._status = status

    @classmethod
    def from_dict(cls, dikt) -> 'Task':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Task of this Task.  # noqa: E501
        :rtype: Task
        """
        return util.deserialize_model(dikt, cls)

    @property
    def description(self) -> TaskDescription:
        """Gets the description of this Task.


        :return: The description of this Task.
        :rtype: TaskDescription
        """
        return self._description

    @description.setter
    def description(self, description: TaskDescription):
        """Sets the description of this Task.


        :param description: The description of this Task.
        :type description: TaskDescription
        """

        self._description = description

    @property
    def status(self) -> str:
        """Gets the status of this Task.


        :return: The status of this Task.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this Task.


        :param status: The status of this Task.
        :type status: str
        """

        self._status = status